Fix bug with entering subsequent expressions
The previous operator and second operand is not cleared after evaluating an expression to allow a user to press "=" multiple times to compound the previous expression without having to re-enter the expression again.
However, when a user enters a number plus an operator directly after evaluating a previous expression, the calculator triggers a premature evaluation to clear space instead of clearing everything away and starting
with a new expression.